+++ frontend/css/styles.css
@@ -4,7 +4,7 @@ body {
 	height: 100%;
 	margin: 0;
 	padding: 0;
-	min-width: 100vh;
+	min-height: 100vh;
 	display: flex;
 	flex-direction: column;
 }
